Move markers conversion to misc submodules to help with deferring matplotlib import
